// MSR.UI/MSR.UI.Answer/src/app/services/mocks/models/createProcedureRequest.ts
/**
 * Parts, Procedures, and Monitors
 * REST API Answer 3
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class CreateProcedureRequest {
    name: string;
    procedureTypeId?: number;
    comments?: string;
    roleIds?: Array<number>;
    duration?: number;
    durationType: string;
    referenceFiles?: Array<any>;
}